move some files and packages around to avoid python conflicts
